requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==2.0.0
cssselect2==0.8.0
Django==4.2.24
django-jazzmin==3.0.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
fonttools==4.60.0
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
xlrd==2.0.2
zopfli==0.2.3.post1
